Shifted OrgSignals to use size_t instead of int

+++ evo/OrgSignals.h
@@ -25,7 +25,7 @@ namespace evo {
   struct OrgSignals_Basic {
     static std::string GetName() { return "OrgSignals_Basic"; }
 
-    Signal<int> repro_sig;
+    Signal<size_t> repro_sig;
     std::string prefix;
 
     OrgSignals_Basic(const std::string & sig_prefix)
@@ -38,7 +38,7 @@ namespace evo {
   struct OrgSignals_Eco : public OrgSignals_Basic {
     static std::string GetName() { return "OrgSignals_Eco"; }
 
-    Signal<int> symbiont_repro_sig;  // Identify organism with symbiont reproducing.
+    Signal<size_t> symbiont_repro_sig;  // Identify organism with symbiont reproducing.
 
     OrgSignals_Eco(const std::string & sig_prefix)
       : OrgSignals_Basic(sig_prefix)
